Modularize code base
