+++ seo_aio_streamlit.py
@@ -7,7 +7,6 @@ from bs4 import BeautifulSoup
 import tldextract
 from datetime import datetime
 from typing import Dict, List, Tuple, Optional
-from dataclasses import dataclass
 from urllib.parse import urlparse
 import io
 import base64
@@ -117,227 +116,9 @@ try:
 except Exception as e:
     print(f"[DEBUG] .envファイル読み込みエラー（無視）: {e}")
 
-# アプリ設定
-APP_VERSION = "3.0.0"
-APP_NAME = "SEO・AIO統合分析ツール"
-
-# グレー基調カラーパレット
-COLOR_PALETTE = {
-    "primary": "#2C2C2C",      # ダークグレー
-    "secondary": "#3C3C3C",    # ミディアムグレー  
-    "accent": "#4A9EFF",       # ブルーアクセント
-    "background": "#E8E8E8",   # ライトグレー背景
-    "surface": "#F5F5F5",      # サーフェス
-    "text_primary": "#1A1A1A", # プライマリテキスト
-    "text_secondary": "#4A4A4A", # セカンダリテキスト
-    "success": "#4CAF50",      # 成功カラー
-    "warning": "#FF9800",      # 警告カラー
-    "error": "#F44336",        # エラーカラー
-    "info": "#2196F3",         # 情報カラー
-    "gold": "#FF8C00",         # オレンジゴールド
-    "dark_blue": "#1565C0"     # ダークブルー
-}
-
-# AIOスコアマッピング（上位8項目と下位8項目に分割）
-AIO_SCORE_MAP_JP_UPPER = {
-    "experience": "経験 (Experience)", 
-    "expertise": "専門性 (Expertise)",
-    "authoritativeness": "権威性 (Authoritativeness)", 
-    "trustworthiness": "信頼性 (Trustworthiness)",
-    "structure": "構造化と整理", 
-    "qa_compatibility": "質問応答適合性",
-    "citation_potential": "AIによる引用可能性", 
-    "multimodal": "マルチモーダル対応"
-}
-
-AIO_SCORE_MAP_JP_LOWER = {
-    "search_intent": "検索意図マッチング", 
-    "personalization": "パーソナライズ可能性",
-    "uniqueness": "情報の独自性", 
-    "completeness": "コンテンツの完全性",
-    "readability": "読みやすさスコア", 
-    "mobile_friendly": "モバイル対応性",
-    "page_speed": "ページ速度", 
-    "metadata": "メタデータ最適化"
-}
-
-AIO_SCORE_MAP_JP = {**AIO_SCORE_MAP_JP_UPPER, **AIO_SCORE_MAP_JP_LOWER}
-
-@dataclass
-class IndustryAnalysis:
-    """業界分析結果"""
-    primary_industry: str
-    secondary_industries: List[str]
-    confidence_score: float
-    industry_keywords: List[str]
-    specialized_terms: List[str]
-    regulatory_indicators: List[str]
-    target_audience_clues: List[str]
-
-class IndustryDetector:
-    """業界自動判定システム"""
-    
-    def __init__(self):
-        # 業界キーワード辞書
-        self.industry_keywords = {
-            "IT・テクノロジー": {
-                "primary": ["API", "SDK", "SaaS", "クラウド", "データベース", "システム開発", "ソフトウェア", "アプリ"],
-                "secondary": ["DX", "デジタル変革", "IT導入", "クラウド移行", "セキュリティ", "AI", "IoT"],
-                "specialized": ["React", "Python", "AWS", "Docker", "kubernetes", "GitHub", "DevOps"]
-            },
-            "医療・ヘルスケア": {
-                "primary": ["診療", "治療", "医師", "看護師", "病院", "クリニック", "薬事法", "医療"],
-                "secondary": ["予防医療", "遠隔診療", "電子カルテ", "医療DX", "健康管理"],
-                "specialized": ["ICD-10", "レセプト", "診療報酬", "薬機法", "PMDA"]
-            },
-            "不動産": {
-                "primary": ["物件", "賃貸", "売買", "マンション", "戸建て", "土地", "不動産投資"],
-                "secondary": ["リノベーション", "住宅ローン", "仲介手数料", "賃貸管理"],
-                "specialized": ["重要事項説明", "宅建士", "建ぺい率", "容積率", "登記"]
-            },
-            "教育・人材": {
-                "primary": ["学習", "教育", "講座", "スクール", "研修", "資格", "eラーニング", "人材"],
-                "secondary": ["オンライン授業", "学習管理", "教材開発", "採用", "転職"],
-                "specialized": ["LMS", "アダプティブラーニング", "学習分析", "HRtech"]
-            },
-            "金融・保険": {
-                "primary": ["融資", "投資", "保険", "資産運用", "金利", "リスク管理", "銀行"],
-                "secondary": ["フィンテック", "ロボアドバイザー", "仮想通貨", "決済"],
-                "specialized": ["AML", "KYC", "Basel III", "Solvency II", "PCI DSS"]
-            },
-            "製造業": {
-                "primary": ["製造", "生産", "工場", "品質管理", "サプライチェーン", "設備"],
-                "secondary": ["IoT", "スマートファクトリー", "予知保全", "自動化"],
-                "specialized": ["QMS", "ISO9001", "TPM", "5S", "カイゼン", "JIT"]
-            },
-            "小売・EC": {
-                "primary": ["商品", "販売", "店舗", "顧客", "在庫", "決済", "配送", "EC"],
-                "secondary": ["オムニチャネル", "CRM", "ポイント", "レコメンド"],
-                "specialized": ["SKU", "GMV", "LTV", "CAC", "CVR", "ROAS"]
-            },
-            "飲食・食品": {
-                "primary": ["メニュー", "レストラン", "食材", "調理", "衛生管理", "栄養"],
-                "secondary": ["テイクアウト", "デリバリー", "食品ロス", "フードテック"],
-                "specialized": ["HACCP", "食品表示法", "トレーサビリティ"]
-            },
-            "建設・建築": {
-                "primary": ["建設", "建築", "施工", "設計", "リフォーム", "住宅"],
-                "secondary": ["BIM", "建築DX", "省エネ", "耐震"],
-                "specialized": ["建築基準法", "一級建築士", "施工管理", "構造計算"]
-            },
-            "コンサルティング": {
-                "primary": ["コンサル", "戦略", "業務改善", "経営", "支援"],
-                "secondary": ["DXコンサル", "ITコンサル", "人事コンサル"],
-                "specialized": ["フレームワーク", "ベストプラクティス", "KPI"]
-            }
-        }
-
-    def analyze_industries(self, title: str, content: str, meta_description: str = "") -> IndustryAnalysis:
-        """複数業界の分析"""
-        combined_text = f"{title} {meta_description} {content}".lower()
-        
-        # 各業界のスコア計算
-        industry_scores = {}
-        matched_keywords = {}
-        
-        for industry, keywords in self.industry_keywords.items():
-            score = 0
-            matched = []
-            
-            # Primary keywords (重み: 3)
-            for keyword in keywords["primary"]:
-                count = combined_text.count(keyword.lower())
-                score += count * 3
-                if count > 0:
-                    matched.append(keyword)
-            
-            # Secondary keywords (重み: 2)
-            for keyword in keywords["secondary"]:
-                count = combined_text.count(keyword.lower())
-                score += count * 2
-                if count > 0:
-                    matched.append(keyword)
-                    
-            # Specialized terms (重み: 5)
-            for keyword in keywords["specialized"]:
-                count = combined_text.count(keyword.lower())
-                score += count * 5
-                if count > 0:
-                    matched.append(keyword)
-            
-            industry_scores[industry] = score
-            matched_keywords[industry] = matched
-        
-        # スコア順でソート
-        sorted_industries = sorted(industry_scores.items(), key=lambda x: x[1], reverse=True)
-        
-        if not sorted_industries or sorted_industries[0][1] == 0:
-            return IndustryAnalysis(
-                primary_industry="指定なし（自動判定不可）",
-                secondary_industries=[],
-                confidence_score=0.0,
-                industry_keywords=[],
-                specialized_terms=[],
-                regulatory_indicators=[],
-                target_audience_clues=[]
-            )
-        
-        # 主要業界と副業界の特定
-        primary_industry = sorted_industries[0][0]
-        primary_score = sorted_industries[0][1]
-        
-        # 副業界は主要業界スコアの30%以上のものを抽出
-        secondary_industries = []
-        threshold = primary_score * 0.3
-        for industry, score in sorted_industries[1:6]:  # 最大5つまで
-            if score >= threshold and score > 0:
-                secondary_industries.append(f"{industry}({score:.0f})")
-        
-        # 信頼度計算
-        total_words = len(combined_text.split())
-        confidence = min(100, (primary_score / max(total_words * 0.1, 1)) * 100)
-        
-        # その他の情報抽出
-        target_clues = self._detect_target_audience(combined_text)
-        regulatory_indicators = self._detect_regulatory_terms(combined_text)
-        
-        return IndustryAnalysis(
-            primary_industry=primary_industry,
-            secondary_industries=secondary_industries,
-            confidence_score=confidence,
-            industry_keywords=matched_keywords[primary_industry],
-            specialized_terms=matched_keywords[primary_industry],
-            regulatory_indicators=regulatory_indicators,
-            target_audience_clues=target_clues
-        )
-    
-    def _detect_target_audience(self, text: str) -> List[str]:
-        """ターゲット層の検出"""
-        audience_patterns = {
-            "法人向け": ["企業", "会社", "法人", "ビジネス", "B2B"],
-            "個人向け": ["個人", "家庭", "一般", "消費者", "B2C"],
-            "専門職向け": ["医師", "弁護士", "税理士", "エンジニア", "専門家"],
-            "経営者向け": ["経営者", "社長", "CEO", "役員", "管理職"]
-        }
-        
-        detected = []
-        for audience_type, patterns in audience_patterns.items():
-            if any(pattern in text for pattern in patterns):
-                detected.append(audience_type)
-        return detected
-    
-    def _detect_regulatory_terms(self, text: str) -> List[str]:
-        """規制関連用語の検出"""
-        regulatory_terms = [
-            "薬機法", "医療法", "金融商品取引法", "宅建業法", "建築基準法",
-            "個人情報保護法", "食品衛生法", "労働基準法", "GDPR", "ISO"
-        ]
-        
-        detected = []
-        for term in regulatory_terms:
-            if term.lower() in text:
-                detected.append(term)
-        return detected
+from core.constants import APP_VERSION, APP_NAME, COLOR_PALETTE, AIO_SCORE_MAP_JP, AIO_SCORE_MAP_JP_UPPER, AIO_SCORE_MAP_JP_LOWER
+from core.industry_detector import IndustryDetector, IndustryAnalysis
+from core.visualization import create_score_gauge, create_aio_score_chart_vertical
 
 class SEOAIOAnalyzer:
     def __init__(self):
@@ -1326,82 +1107,6 @@ def set_custom_css():
     </style>
     """, unsafe_allow_html=True)
 
-def create_score_gauge(score, title, color):
-    """スコアゲージの作成"""
-    fig = go.Figure(go.Indicator(
-        mode = "gauge+number",
-        value = score,
-        domain = {'x': [0, 1], 'y': [0, 1]},
-        title = {'text': title, 'font': {'color': COLOR_PALETTE["text_primary"]}},
-        gauge = {
-            'axis': {'range': [None, 100], 'tickcolor': COLOR_PALETTE["text_secondary"]},
-            'bar': {'color': color},
-            'steps': [
-                {'range': [0, 50], 'color': COLOR_PALETTE["secondary"]},
-                {'range': [50, 80], 'color': COLOR_PALETTE["surface"]},
-                {'range': [80, 100], 'color': COLOR_PALETTE["primary"]}
-            ],
-            'threshold': {
-                'line': {'color': COLOR_PALETTE["gold"], 'width': 4},
-                'thickness': 0.75,
-                'value': 90
-            }
-        }
-    ))
-    
-    fig.update_layout(
-        paper_bgcolor=COLOR_PALETTE["background"],
-        plot_bgcolor=COLOR_PALETTE["background"],
-        font={'color': COLOR_PALETTE["text_primary"]},
-        height=300
-    )
-    
-    return fig
-
-def create_aio_score_chart_vertical(scores_data, score_map, title):
-    """AIOスコア縦型チャート作成"""
-    labels = [score_map.get(k, k.title()) for k in score_map.keys()]
-    values = [scores_data.get(k, {"score": 0}).get("score", 0) for k in score_map.keys()]
-    
-    # カラーマッピング
-    colors = []
-    for value in values:
-        if value >= 8:
-            colors.append(COLOR_PALETTE["success"])
-        elif value >= 6:
-            colors.append(COLOR_PALETTE["info"])
-        elif value >= 4:
-            colors.append(COLOR_PALETTE["warning"])
-        else:
-            colors.append(COLOR_PALETTE["error"])
-    
-    fig = go.Figure(data=[go.Bar(
-        y=labels,
-        x=values,
-        orientation='h',
-        marker=dict(color=colors),
-        text=[f'{v:.1f}' for v in values],
-        textposition='outside'
-    )])
-    
-    fig.update_layout(
-        title=title,
-        title_font_color=COLOR_PALETTE["dark_blue"],
-        paper_bgcolor=COLOR_PALETTE["background"],
-        plot_bgcolor=COLOR_PALETTE["background"],
-        font={'color': COLOR_PALETTE["text_primary"]},
-        xaxis=dict(
-            range=[0, 10],
-            gridcolor=COLOR_PALETTE["secondary"],
-            title="スコア (/10)"
-        ),
-        yaxis=dict(
-            gridcolor=COLOR_PALETTE["secondary"]
-        ),
-        height=600
-    )
-    
-    return fig
 
 def main():
     # ページ設定
